pyproject.toml
[tool.poetry]
name = "deqr"
version = "0.0.0"
description = "qr code decoding library"
authors = ["torque <torque@users.noreply.github.com>"]
license = "LGPL-2.1-or-later"
readme = "README.md"
repository = "https://github.com/torque/deqr"
documentation = "https://torque.github.io/deqr-docs/"
build = "build.py"
include = [
    "deps/qrdec/src",
    "deps/quirc/lib",
]

[tool.poetry.dependencies]
python = "^3.8"

[tool.poetry.dev-dependencies]
Cython = {version = "3.0a7", allow-prereleases = true}
poetry-dynamic-versioning = "^0.13.0"

# tests
pytest = "^6.0.0"
Pillow = "^8.2.0"
numpy = "^1.20.3"

# documentation
Sphinx = "^4.0.0"
sphinx-copybutton = "^0.3.3"
furo = "^2021.6.18-beta.36"
sphinx-inline-tabs = "^2021.4.11-beta.9"

[tool.poetry-dynamic-versioning]
enable = true
vcs = "git"
pattern = "^(?P<base>\\d+\\.\\d+\\.\\d+)"
style = "semver"
latest-tag = true

[tool.poetry-dynamic-versioning.substitution]
files = ["deqr/__init__.py"]
patterns = ["(^__version__\\s+=\\s+['\"])__PLACEHOLDER__(['\"])"]

[build-system]
requires = [
    "cython==3.0a7",
    "poetry-core>=1.0.0",
    "poetry-dynamic-versioning>=0.13.0",
    "setuptools>=40.9.0",
    "wheel",
]
build-backend = "poetry.core.masonry.api"
